First attempt to introduce a document compiler.
The compiler is meant to be an post processor when the parser is finished,
we could see this as part of the parsing part depending on what you expect from
an parser.

From my point of view the parsing is a first step of a traditional compiler that
has multiple interations to be able to get into the final result, in our case an ast
ready to be transformed by the renderen into html.

The parser nodes are cachable, the compilter will transform them into a new structure that
cannot be cached, as it has inter document linkes that should be updated once a document is
changed. For example the titles in a toc-tree.